+++ src/components/common/query/sidebar/AppliedFilterList.jsx
