+++ src/components/common/query/sidebar/panels/InfoView.jsx
